[nats configuration] rename module
