+++ rhio/src/utils/mod.rs
@@ -1,2 +1,2 @@
 pub mod nats;
-pub mod retry_source;
+pub mod retry;
